requirements.txt
streamlit
PyMuPDF
python-docx
pandas
openpyxl